+++ src/ext/java/org/opentripplanner/ext/accessibilityscore/AccessibilityScoreFilter.java
@@ -37,7 +37,7 @@ public record AccessibilityScoreFilter(double wheelchairMaxSlope) implements Iti
 
   public static float compute(ScheduledTransitLeg leg) {
     var fromStop = leg.getFrom().stop.getWheelchairAccessibility();
-    var toStop = leg.getFrom().stop.getWheelchairAccessibility();
+    var toStop = leg.getTo().stop.getWheelchairAccessibility();
     var trip = leg.getTripWheelchairAccessibility();
 
     var values = List.of(trip, fromStop, toStop);