+++ src/ext/java/org/opentripplanner/ext/restapi/resources/RoutingResource.java
@@ -20,14 +20,15 @@ import javax.xml.datatype.DatatypeFactory;
 import javax.xml.datatype.XMLGregorianCalendar;
 import org.opentripplanner.api.parameter.QualifiedModeSet;
 import org.opentripplanner.ext.dataoverlay.api.DataOverlayParameters;
+import org.opentripplanner.ext.restapi.mapping.LegacyBicycleOptimizeType;
 import org.opentripplanner.framework.application.OTPFeature;
 import org.opentripplanner.framework.lang.StringUtils;
 import org.opentripplanner.framework.time.DurationUtils;
+import org.opentripplanner.framework.time.ZoneIdFallback;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.routing.api.request.preference.ItineraryFilterDebugProfile;
 import org.opentripplanner.routing.api.request.request.filter.SelectRequest;
 import org.opentripplanner.routing.api.request.request.filter.TransitFilterRequest;
-import org.opentripplanner.routing.core.BicycleOptimizeType;
 import org.opentripplanner.standalone.api.OtpServerRequestContext;
 import org.opentripplanner.standalone.config.framework.file.ConfigFileLoader;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
@@ -248,7 +249,7 @@ public abstract class RoutingResource {
    */
   @Deprecated
   @QueryParam("optimize")
-  protected BicycleOptimizeType bikeOptimizeType;
+  protected LegacyBicycleOptimizeType bikeOptimizeType;
 
   /**
    * The set of modes that a user is willing to use, with qualifiers stating whether vehicles should
@@ -658,8 +659,8 @@ public abstract class RoutingResource {
   @QueryParam("useVehicleParkingAvailabilityInformation")
   protected Boolean useVehicleParkingAvailabilityInformation;
 
-  @QueryParam("relaxTransitPriorityGroup")
-  protected String relaxTransitPriorityGroup;
+  @QueryParam("relaxTransitGroupPriority")
+  protected String relaxTransitGroupPriority;
 
   /**
    * Whether non-optimal transit paths at the destination should be returned.
@@ -711,7 +712,7 @@ public abstract class RoutingResource {
 
     {
       //FIXME: move into setter method on routing request
-      ZoneId tz = serverContext.transitService().getTimeZone();
+      ZoneId tz = ZoneIdFallback.zoneId(serverContext.transitService().getTimeZone());
       if (date == null && time != null) { // Time was provided but not date
         LOG.debug("parsing ISO datetime {}", time);
         try {